+++ src/dynamixel_diagnostic_controller.cpp
@@ -21,7 +21,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <dynamixel_hardware_interface/dymanixel_diagnostic_controller.hpp>
+#include <dynamixel_hardware_interface/dynamixel_diagnostic_controller.hpp>
 
 namespace dynamixel_hardware_interface
 {
